refactor: move out act and learn functions
